:cherry_blossom: Match wireframes for flight segment display.
Sadly this doesn't work at all well on mobile, so I might rework it.
